test setup decided and started construction
